Rename catalog_new to catalogNew
