+++ assets/js/state/selectors/catalog.js
@@ -1,7 +1,7 @@
 export const getCatalog = () => (state) => {
   return {
-    data: state.catalog_new.data,
-    error: state.catalog_new.error,
-    loading: state.catalog_new.loading,
+    data: state.catalogNew.data,
+    error: state.catalogNew.error,
+    loading: state.catalogNew.loading,
   };
 };
