+++ assets/js/state/selectors/catalog.test.js
@@ -3,7 +3,7 @@ import { getCatalog } from './catalog';
 describe('Catalog selector', () => {
   it('should return the correct catalog state', () => {
     const state = {
-      catalog_new: {
+      catalogNew: {
         loading: false,
         data: [1, 2, 3],
         error: null,
